modified delete orderlist

+++ routes/controller/park.controller.js
@@ -1,14 +1,14 @@
-const jwt = require('jsonwebtoken');
-const User = require('../../models/User');
-const Seat = require('../../models/Seat');
-const aws = require('aws-sdk');
-const multer = require('multer');
-const multerS3 = require('multer-s3');
+const jwt = require("jsonwebtoken");
+const User = require("../../models/User");
+const Seat = require("../../models/Seat");
+const aws = require("aws-sdk");
+const multer = require("multer");
+const multerS3 = require("multer-s3");
 
 aws.config.update({
   accessKeyId: process.env.AWS_ACCESS_KEY_ID,
   secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-  region: 'ap-northeast-2'
+  region: "ap-northeast-2"
 });
 
 const s3 = new aws.S3();
@@ -16,9 +16,9 @@ const s3 = new aws.S3();
 exports.uploadImage = multer({
   storage: multerS3({
     s3: s3,
-    bucket: 'nkhvc',
+    bucket: "nkhvc",
     contentType: multerS3.AUTO_CONTENT_TYPE,
-    acl: 'public-read',
+    acl: "public-read",
     key: (req, file, cb) => {
       cb(null, file.originalname);
     }
@@ -27,15 +27,19 @@ exports.uploadImage = multer({
 
 exports.deleteImage = async (req, res, next) => {
   const data = await Seat.find()
-    .where('park')
+    .where("park")
     .equals(req.body.park)
-    .where('seller')
+    .where("seller")
     .equals(req.params.id);
-  const urlKey = await data[0].image_url.split('/').pop();
+
+  if (!data.length) {
+    return;
+  }
+  const urlKey = await data[0].image_url.split("/").pop();
 
   s3.deleteObject(
     {
-      Bucket: 'nkhvc',
+      Bucket: "nkhvc",
       Key: urlKey
     },
     function(err, data) {
@@ -48,9 +52,9 @@ exports.deleteImage = async (req, res, next) => {
 
 exports.deleteOrderList = async (req, res, next) => {
   const result = await Seat.remove()
-    .where('park')
+    .where("park")
     .equals(req.body.park)
-    .where('seller')
+    .where("seller")
     .equals(req.params.id);
   res.send({ result: result });
 };
@@ -62,22 +66,24 @@ exports.sendFileLocation = (req, res, next) => {
 exports.saveExchangeData = async (req, res, next) => {
   const newExchange = new Seat(req.body.data);
   const data = await newExchange.save();
-  res.send({ result: req.body.data, newData : data });
+  res.send({ result: req.body.data, newData: data });
 };
 
 exports.searchParkList = async (req, res, next) => {
   const parkList = await Seat.find()
-    .where('park')
+    .where("park")
     .equals(req.params.id)
-    .where('complete')
-    .ne('true');
-    console.log(parkList);
-  res.send({ result: 'ok', parkList: parkList });
+    .where("complete")
+    .ne("true");
+  console.log(parkList);
+  res.send({ result: "ok", parkList: parkList });
 };
 
 exports.changeExchangeStatus = async (req, res, next) => {
-  const order = await Seat.findOne().where('_id').equals(req.params.id);
+  const order = await Seat.findOne()
+    .where("_id")
+    .equals(req.params.id);
   order.complete = req.body.status;
   order.save();
-  res.send({result : 'ok'});
-}
+  res.send({ result: "ok" });
+};
